Improve missing python error in build.rs
The build.rs file requires python3 to execute, but when it is missing it
just prints a "No such file or directory" error.

To make this dependency error more clear, an `expect` has been added
that informs the user that the build dependency python3 is missing.

This fixes rtbo/rust-xcb#49.