Add Event::as_raw(), Connection::resolve_event()
In some ffi scenarios it is necessary to extract the underlying
event pointer.  The `Event::as_raw` method serves this purpose.

I couldn't implement the `Raw<>` trait for `Event` because the
`from_raw` method requires additional context in order to perform
the inverse of the `as_raw` method, so `Event::as_raw` is just
a standalone method.

I added `Connection::resolve_event` as the inverse operation
to give some symmetry.

closes: https://github.com/rust-x-bindings/rust-xcb/issues/188