Expose "format" property getter
There's a preexisting whitelist of properties/fields that should not be
hidden just because they appear as a part of another expression (most
notably, lists) as a <fieldref> expression.

For the client code it's important to check the format, otherwise the
code might panic due to the assert inside .data::<PropEl>() accessor.

Fixes #209