Expose "keycodes_per_modifier" property getter
The GetModifierMapping request returns a list of keycodes. Client code
needs to iterate this list in chunks of keycodes_per_modifier to visit
all keycodes for each of the 8 modifiers (Shift, Lock, Control, Mod1,
Mod2, Mod3, Mod4, Mod5).

The preexisting whitelist of properties/fields that should not be
hidden has been extended to include this field.

Fixes #271